1.10.10 Bugfix: Fix backup progressbar

+++ src/main/de/jClipCorn/util/listener/ProgressCallbackListener.java
@@ -2,9 +2,9 @@ package de.jClipCorn.util.listener;
 
 public interface ProgressCallbackListener {
 	void step();
-	void step(final int inc);
+	void step(final long inc);
 	void stepToMax();
-	void setMax(final int max);
+	void setMax(final long max);
 	void reset();
 
 }
