+++ src/main/de/jClipCorn/util/listener/ProgressCallbackSink.java
@@ -8,13 +8,13 @@ public class ProgressCallbackSink implements ProgressCallbackListener{
 	public void step() { /* NOP */ }
 
 	@Override
-	public void step(final int inc) { /* NOP */ }
+	public void step(final long inc) { /* NOP */ }
 
 	@Override
 	public void stepToMax() { /* NOP */ }
 
 	@Override
-	public void setMax(final int max) { /* NOP */ }
+	public void setMax(final long max) { /* NOP */ }
 
 	@Override
 	public void reset() { /* NOP */ }
